requirements.txt
google-generativeai>=0.3.0
gtts>=2.4.0
moviepy>=1.0.3
requests>=2.31.0
pillow>=10.0.0